requirements.txt
langchain,
langchain-core,
langchain-groq
pillow
python-dotenv
streamlit
torch
torchvision
transformers